refactor(sdk)!: move worker-connection-manager types to the helpers submodule (#1840)
