+++ sdk/packages/rust/iii/src/builtin_triggers.rs
@@ -1,5 +1,6 @@
 use std::collections::HashMap;
 
+use iii_helpers::stream::{StreamJoinLeaveTriggerConfig, StreamTriggerConfig};
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
@@ -192,96 +193,6 @@ impl Default for StateTriggerConfig {
     }
 }
 
-// ── Stream ──────────────────────────────────────────────────────────────
-
-#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
-pub struct StreamJoinLeaveTriggerConfig {
-    /// Stream name to watch
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub stream_name: Option<String>,
-    /// Optional function ID to evaluate before invoking handler
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub condition_function_id: Option<String>,
-}
-
-impl StreamJoinLeaveTriggerConfig {
-    pub fn new() -> Self {
-        Self {
-            stream_name: None,
-            condition_function_id: None,
-        }
-    }
-
-    pub fn stream_name(mut self, name: impl Into<String>) -> Self {
-        self.stream_name = Some(name.into());
-        self
-    }
-
-    pub fn condition(mut self, function_id: impl Into<String>) -> Self {
-        self.condition_function_id = Some(function_id.into());
-        self
-    }
-}
-
-impl Default for StreamJoinLeaveTriggerConfig {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
-pub struct StreamTriggerConfig {
-    /// Stream name to watch
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub stream_name: Option<String>,
-    /// Group ID filter
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub group_id: Option<String>,
-    /// Item ID filter
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub item_id: Option<String>,
-    /// Optional function ID to evaluate before invoking handler
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub condition_function_id: Option<String>,
-}
-
-impl StreamTriggerConfig {
-    pub fn new() -> Self {
-        Self {
-            stream_name: None,
-            group_id: None,
-            item_id: None,
-            condition_function_id: None,
-        }
-    }
-
-    pub fn stream_name(mut self, name: impl Into<String>) -> Self {
-        self.stream_name = Some(name.into());
-        self
-    }
-
-    pub fn group_id(mut self, id: impl Into<String>) -> Self {
-        self.group_id = Some(id.into());
-        self
-    }
-
-    pub fn item_id(mut self, id: impl Into<String>) -> Self {
-        self.item_id = Some(id.into());
-        self
-    }
-
-    pub fn condition(mut self, function_id: impl Into<String>) -> Self {
-        self.condition_function_id = Some(function_id.into());
-        self
-    }
-}
-
-impl Default for StreamTriggerConfig {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
 // ── Log ─────────────────────────────────────────────────────────────────
 
 #[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
@@ -347,55 +258,6 @@ pub struct StateCallRequest {
     pub key: String,
     pub old_value: Option<Value>,
     pub new_value: Value,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
-pub struct StreamJoinLeaveEvent {
-    pub subscription_id: String,
-    pub stream_name: String,
-    pub group_id: String,
-    pub id: Option<String>,
-    pub context: Option<Value>,
-}
-
-/// The kind of mutation that occurred on a stream item.
-#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema, PartialEq, Eq)]
-#[serde(rename_all = "lowercase")]
-pub enum StreamEventType {
-    Create,
-    Update,
-    Delete,
-}
-
-/// Detail of a stream change event containing the mutation type and data.
-#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
-pub struct StreamEventDetail {
-    /// The kind of mutation (create, update, or delete).
-    #[serde(rename = "type")]
-    pub event_type: StreamEventType,
-    /// The data associated with the event.
-    pub data: Value,
-}
-
-/// Handler input for `stream` triggers, fired when an item changes
-/// via `stream::set`, `stream::update`, or `stream::delete`.
-#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
-pub struct StreamChangeEvent {
-    /// Always `"stream"`.
-    #[serde(rename = "type")]
-    pub event_type: String,
-    /// Unix timestamp (milliseconds) of the event.
-    pub timestamp: i64,
-    /// The stream where the change occurred.
-    #[serde(rename = "streamName")]
-    pub stream_name: String,
-    /// The group where the change occurred.
-    #[serde(rename = "groupId")]
-    pub group_id: String,
-    /// The item ID that changed.
-    pub id: Option<String>,
-    /// The event detail containing mutation type and data.
-    pub event: StreamEventDetail,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
@@ -516,43 +378,6 @@ mod tests {
         let result = QueueTriggerConfig::new("emails").queue_config(FailingSerialize);
 
         assert!(result.is_err(), "serialization failures should be returned");
-    }
-
-    #[test]
-    fn stream_call_request_deserializes_typed_event() {
-        let request: StreamChangeEvent = serde_json::from_value(json!({
-            "type": "stream",
-            "timestamp": 1700000000000_i64,
-            "streamName": "chat",
-            "groupId": "room-1",
-            "id": "msg-42",
-            "event": {
-                "type": "create",
-                "data": { "text": "hello" }
-            }
-        }))
-        .expect("stream call request should deserialize");
-
-        assert_eq!(request.event_type, "stream");
-        assert_eq!(request.stream_name, "chat");
-        assert_eq!(request.group_id, "room-1");
-        assert_eq!(request.id.as_deref(), Some("msg-42"));
-        assert!(matches!(request.event.event_type, StreamEventType::Create));
-        assert_eq!(request.event.data, json!({ "text": "hello" }));
-    }
-
-    #[test]
-    fn stream_event_type_roundtrip() {
-        for (variant, expected) in [
-            (StreamEventType::Create, "create"),
-            (StreamEventType::Update, "update"),
-            (StreamEventType::Delete, "delete"),
-        ] {
-            let json = serde_json::to_value(&variant).unwrap();
-            assert_eq!(json, expected);
-            let back: StreamEventType = serde_json::from_value(json).unwrap();
-            assert_eq!(back, variant);
-        }
     }
 
     #[test]
